+++ plugins/as/CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## Unreleased
+
+### Fixed
+
+- Correct the scheduled scaling-policy name label so it does not describe a workflow.
+
 ## 0.1.0-beta.3 - 2026-09-23
 
 ### Added